getdents_calls

+++ src/trusted/service_runtime/nacl_syscall_strace.h
@@ -12,7 +12,7 @@ void NaClStraceThreadExit(int stack_flag,uint32_t  zero);
 void NaClStraceDup(int oldfd,int ret);
 void NaClStraceDup2(int oldfd,int newfd,int ret);
 void NaClStraceDup3(int oldfd,int newfd,int flags,int ret);
-void NaClStraceGetdents(int d,void *drip,int  count,size_t ret,ssize_t getdents_ret,uintptr_t sysaddr);
+void NaClStraceGetdents(int d, void *drip, size_t count, size_t ret, ssize_t getdents_ret, uintptr_t sysaddr);
 void NaClStracePread(int d, void *buf, int count,  size_t log_bytes,int32_t ret);
 void NaClStraceWrite(int d, void *buf, int count, size_t ret);
 void NaClStracePWrite(int d, void *buf, int count,off_t offset, size_t ret);
